fixed issues about launching with cmdenv

+++ PECSN_project/src/User.cc
@@ -51,7 +51,7 @@ void User::handleMessage(cMessage *msg){
     Frame *frame = check_and_cast<Frame*>(msg);
     std::vector<Packet*> pl = frame->getPacketList();
     EV<<"total size:\t"<<pl.size()<<endl;
-    double bytesReceived = 0;
+    int bytesReceived = 0;
     int pR = 0;
     while(pl.size() != 0){
         Packet * p = pl.back();
